[Server] Removed NotifyFilters.LastAccess from the file watcher
